♻️ Move seat-update portal flow into apps/web billing feature
The seat-update flow (config creation, reuse memoization, and session
creation) now lives in apps/web/src/features/billing/portal.ts as
application code, using absoluteUrl directly.

@rallly/billing/lib/portal is reduced to the shared identity markers
(headline, purpose, version) — the contract that both the app flow and the
cleanup script depend on, kept single-sourced in the package so the
cleanup script (which can't import from apps/web) stays in lockstep on
version bumps.
