+++ apps/web/src/features/billing/portal.ts
@@ -1,0 +1,126 @@
+import { getProPricing, stripe } from "@rallly/billing";
+import {
+  SEAT_UPDATE_PORTAL_HEADLINE,
+  SEAT_UPDATE_PORTAL_PURPOSE,
+  SEAT_UPDATE_PORTAL_VERSION,
+} from "@rallly/billing/lib/portal";
+import { absoluteUrl } from "@rallly/utils/absolute-url";
+
+async function createSeatUpdateBillingConfig() {
+  const pricing = await getProPricing();
+
+  // Both monthly and yearly prices share the same product.
+  const monthlyPrice = await stripe.prices.retrieve(pricing.monthly.id);
+  const productId = monthlyPrice.product as string;
+
+  return stripe.billingPortal.configurations.create({
+    business_profile: {
+      headline: SEAT_UPDATE_PORTAL_HEADLINE,
+    },
+    features: {
+      subscription_update: {
+        enabled: true,
+        default_allowed_updates: ["price", "quantity"],
+        // Invoice prorations immediately so seat additions are charged right away
+        // rather than deferred onto the next renewal invoice.
+        proration_behavior: "always_invoice",
+        products: [
+          {
+            product: productId,
+            prices: [pricing.monthly.id, pricing.yearly.id],
+          },
+        ],
+      },
+      subscription_cancel: {
+        enabled: false,
+      },
+      payment_method_update: {
+        enabled: true,
+      },
+      invoice_history: {
+        enabled: true,
+      },
+    },
+    metadata: {
+      purpose: SEAT_UPDATE_PORTAL_PURPOSE,
+      version: SEAT_UPDATE_PORTAL_VERSION,
+    },
+  });
+}
+
+async function resolveConfigurationId(): Promise<string> {
+  const existing = await stripe.billingPortal.configurations.list({
+    active: true,
+    limit: 100,
+  });
+
+  const match = existing.data.find(
+    (config) =>
+      config.metadata?.purpose === SEAT_UPDATE_PORTAL_PURPOSE &&
+      config.metadata?.version === SEAT_UPDATE_PORTAL_VERSION,
+  );
+  if (match) return match.id;
+
+  const created = await createSeatUpdateBillingConfig();
+  return created.id;
+}
+
+// Memoised per server process. Dedupes concurrent callers and survives across
+// warm invocations; resets on failure so the next call retries.
+let configurationIdPromise: Promise<string> | undefined;
+
+function getSeatUpdatePortalConfigurationId(): Promise<string> {
+  if (!configurationIdPromise) {
+    configurationIdPromise = resolveConfigurationId().catch((error) => {
+      configurationIdPromise = undefined;
+      throw error;
+    });
+  }
+  return configurationIdPromise;
+}
+
+/**
+ * Creates a billing portal session deep-linked to the seat-update confirmation
+ * screen for a specific quantity change, reusing the shared configuration.
+ */
+export async function createStripeSubscriptionUpdateConfirmation({
+  customerId,
+  subscriptionId,
+  subscriptionItemId,
+  newSeatCount,
+}: {
+  customerId: string;
+  subscriptionId: string;
+  subscriptionItemId: string;
+  newSeatCount: number;
+}) {
+  const configurationId = await getSeatUpdatePortalConfigurationId();
+
+  const portalSession = await stripe.billingPortal.sessions.create({
+    customer: customerId,
+    configuration: configurationId,
+    return_url: absoluteUrl("/settings/billing"),
+    flow_data: {
+      type: "subscription_update_confirm",
+      subscription_update_confirm: {
+        subscription: subscriptionId,
+        items: [
+          {
+            id: subscriptionItemId,
+            quantity: newSeatCount,
+          },
+        ],
+      },
+      after_completion: {
+        type: "redirect",
+        redirect: {
+          return_url: absoluteUrl("/settings/billing", {
+            seats_updated: "true",
+          }),
+        },
+      },
+    },
+  });
+
+  return portalSession.url;
+}
